// src/domain/workflow/enums/workflow-type.enum.ts
export enum WorkflowType {
  PRE_ARRIVAL_WELCOME = 'PRE_ARRIVAL_WELCOME',
  CHECK_IN_NOTIFICATION = 'CHECK_IN_NOTIFICATION',
  POST_CHECKOUT_REVIEW_REQUEST = 'POST_CHECKOUT_REVIEW_REQUEST',
  RETURNING_GUEST_FLAG = 'RETURNING_GUEST_FLAG',
  VIP_ALERT = 'VIP_ALERT',
  HOUSEKEEPING_TASK_ON_CHECKOUT = 'HOUSEKEEPING_TASK_ON_CHECKOUT',
  CANCELLATION_FOLLOWUP = 'CANCELLATION_FOLLOWUP',
}